Share the branding serialize/require logic with metadata for icons
