+++ test/webpack/stripes-branding-plugin.spec.js
@@ -40,43 +40,11 @@ describe('The stripes-branding-plugin', function () {
   });
 
   describe('apply method', function () {
-    it('assigns serialized branding to this.serializedBranding', function () {
-      const sut = new StripesBrandingPlugin();
-      sut.apply(compilerStub);
-      expect(sut.serializedBranding).to.be.a('string')
-        .which.includes(`"${defaultBranding.logo.alt}"`);
-    });
-
     it('registers the "after-plugins" hook', function () {
       this.sandbox.spy(compilerStub, 'plugin');
       const sut = new StripesBrandingPlugin();
       sut.apply(compilerStub);
       expect(compilerStub.plugin).to.be.calledWith('after-plugins');
-    });
-  });
-
-  describe('_serializeBranding method', function () {
-    it('maintains absolute src paths', function () {
-      const result = StripesBrandingPlugin._serializeBranding(defaultBranding);
-      expect(result).to.be.a('string')
-        .which.includes(`'${defaultBranding.logo.src}'`);
-    });
-
-    it('updates custom src paths', function () {
-      const result = StripesBrandingPlugin._serializeBranding(tenantBranding);
-      expect(result).to.be.a('string')
-        .which.includes(`'.${tenantBranding.logo.src}'`)
-        .and.not.include(`'${tenantBranding.logo.src}'`);
-    });
-
-    it('wraps src values in require()', function () {
-      const result = StripesBrandingPlugin._serializeBranding(tenantBranding);
-      expect(result).to.include('"src": require(');
-    });
-
-    it('does not wrap non-src values in require()', function () {
-      const result = StripesBrandingPlugin._serializeBranding(tenantBranding);
-      expect(result).to.not.include('"alt": require(');
     });
   });
