+++ webpack/stripes-branding-plugin.js
@@ -12,38 +12,8 @@ module.exports = class StripesBrandingPlugin {
   }
 
   apply(compiler) {
-    // The serialized output will be picked up by stripes-config-plugin
-    this.serializedBranding = StripesBrandingPlugin._serializeBranding(this.branding);
-
     // Locate the HtmlWebpackPlugin and apply the favicon.
     compiler.plugin('after-plugins', theCompiler => this._replaceFavicon(theCompiler));
-  }
-
-  // Serialize the branding config.
-  static _serializeBranding(branding) {
-    const assetPath = (thePath) => {
-      if (path.isAbsolute(thePath)) {
-        return thePath;
-      }
-
-      return path.join('..', thePath); // Look outside the node_modules directory
-    };
-
-    // Wraps image paths with require()'s for webpack to process via its file loaders
-    // The require()'s are just strings here so we don't attempt to invoke them right now
-    const injectRequire = (key, value) => {
-      if (key === 'src') {
-        return `require('${assetPath(value)}')`;
-      }
-
-      return value;
-    };
-    // Serialize whole branding configuration, adding require()'s as needed
-    const brandingString = JSON.stringify(branding, injectRequire, 2);
-
-    // Now that the branding object has been serialized, this regex omits the wrapping quotes
-    // surrounding the require()'s so they are invoked when Webpack loads stripes-config
-    return brandingString.replace(/"(require\([^)]+\))"/g, (match, $1) => $1);
   }
 
   // Prep favicon path for use with HtmlWebpackPlugin
